Addition of last base materials and meshes, weapon base enablement
The weapon bases don't shut down on pickup yet, because that functionality
depends on the pickups to signalise that they have been picked up. This is
also not yet tested in netplay, and is quite important, so probably best
not to run a server on this code just yet.

Also adjusted base pre-pivots so they wouldn't be floating above ground.

Signed-off-by: Dainius "GreatEmerald" Masiliunas <pastas4@gmail.com>