Removal of include directories
This means we no longer embed any non-code materials into the compiled .u
file (with the exception being the SPMA smoke texture, because I have no
clue where that one came from and which pacakge it's supposed to be in).

All the materials that are no longer used are now moved to raw, and may be
removed at any time.

Signed-off-by: GreatEmerald <pastas4@gmail.com>