AVRiL: Sound Updates
Tried to fix the no weapon fire animation issue but don't know how.
Only sound missing from this now I think is the smoke trail travel sound but it may not be needed. 
I got the sound to work but ended up with never ending smoke at the impact points.